Subdivide cutcode into individual classes
